# docs/openapi.json
{
  "components" : {
    "schemas" : {
...
      "OtpRequest" : {
        "properties" : {
          "phone" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "phone" ],
        "type" : "object"
      },
      "OtpRequested" : {
        "properties" : {
          "expiresInSeconds" : {
            "format" : "int64",
            "type" : "integer"
          },
          "resendAfterSeconds" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "OtpVerify" : {
        "properties" : {
          "code" : {
            "minLength" : 1,
            "type" : "string"
          },
          "phone" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "code", "phone" ],
        "type" : "object"
      },
...
      "Problem" : {
        "additionalProperties" : true,
        "description" : "RFC 7807 problem details. Extra members (e.g. `missing`, `errors`) depend on `code`.",
        "properties" : {
          "code" : {
            "description" : "Stable error code for clients, e.g. `driver.documents-missing`",
            "example" : "driver.documents-missing"
          },
          "detail" : {
            "example" : "Some required documents have not been uploaded."
          },
          "instance" : { },
          "status" : {
            "example" : 409
          },
          "title" : {
            "example" : "Conflict"
          },
          "type" : {
            "example" : "about:blank"
          }
        }
      },
...
      "RefreshRequest" : {
        "properties" : {
          "refreshToken" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "refreshToken" ],
        "type" : "object"
      },
...
      "TokenResponse" : {
        "properties" : {
          "accessToken" : {
            "type" : "string"
          },
          "expiresIn" : {
            "format" : "int64",
            "type" : "integer"
          },
          "refreshExpiresIn" : {
            "format" : "int64",
            "type" : "integer"
          },
          "refreshToken" : {
            "type" : "string"
          },
          "tokenType" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "bearerFormat" : "JWT",
        "description" : "Access token from /api/v1/auth/otp/verify or /api/v1/auth/refresh",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Scheduled & contracted mobility platform for Phnom Penh.\n\n* **Authentication:** phone OTP (`/api/v1/auth/otp/request` → `/otp/verify`) returns a 15-minute JWT access token. Send it as `Authorization: Bearer <token>`. Refresh with `/api/v1/auth/refresh`.\n* **Retries:** mutating endpoints accept an `Idempotency-Key` header; a retry with the same key and body replays the original response (`Idempotent-Replayed: true`).\n* **Errors:** RFC 7807 `application/problem+json` with a stable machine-readable `code`; `title` and `detail` are localized from `Accept-Language` (`km` default, `en`).\n* **Time:** timestamps are UTC (ISO-8601); business dates are Asia/Phnom_Penh.",
    "title" : "Mobility API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/auth/logout" : {
      "post" : {
        "description" : "Revokes the refresh token's session. Unknown tokens are ignored.",
        "operationId" : "logout",
        "parameters" : [ {
          "description" : "Language of error messages and SMS: `km` (default) or `en`",
          "in" : "header",
          "name" : "Accept-Language",
          "required" : false,
          "schema" : {
            "enum" : [ "km", "en" ],
            "type" : "string"
          }
        }, {
          "description" : "Optional, 1–255 chars (a UUID is ideal). Retrying with the same key and body replays the first response with `Idempotent-Replayed: true`; a different body returns 422.",
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : false,
          "schema" : {
            "maxLength" : 255,
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Invalid request (`code` explains why)"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "`idempotency.key-reused`: same Idempotency-Key, different body"
          }
        },
        "summary" : "End the session",
        "tags" : [ "Auth" ]
      }
    },
    "/api/v1/auth/otp/request" : {
      "post" : {
        "description" : "Accepts Cambodian local formats (`012 345 678`) or E.164. The code is valid for 5 minutes.",
        "operationId" : "requestOtp",
        "parameters" : [ {
          "description" : "Language of error messages and SMS: `km` (default) or `en`",
          "in" : "header",
          "name" : "Accept-Language",
          "required" : false,
          "schema" : {
            "enum" : [ "km", "en" ],
            "type" : "string"
          }
        }, {
          "description" : "Optional, 1–255 chars (a UUID is ideal). Retrying with the same key and body replays the first response with `Idempotent-Replayed: true`; a different body returns 422.",
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : false,
          "schema" : {
            "maxLength" : 255,
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OtpRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OtpRequested"
                }
              }
            },
            "description" : "Accepted"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Invalid request (`code` explains why)"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "`idempotency.key-reused`: same Idempotency-Key, different body"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "`otp.resend-cooldown` (60 s between codes) or `otp.hourly-limit` (5 per hour); see `Retry-After`"
          }
        },
        "summary" : "Send a login code by SMS",
        "tags" : [ "Auth" ]
      }
    },
    "/api/v1/auth/otp/verify" : {
      "post" : {
        "description" : "Unknown phone numbers are registered as passengers. Returns a 15-minute access token and a 30-day refresh token.",
        "operationId" : "verifyOtp",
        "parameters" : [ {
          "description" : "Language of error messages and SMS: `km` (default) or `en`",
          "in" : "header",
          "name" : "Accept-Language",
          "required" : false,
          "schema" : {
            "enum" : [ "km", "en" ],
            "type" : "string"
          }
        }, {
          "description" : "Optional, 1–255 chars (a UUID is ideal). Retrying with the same key and body replays the first response with `Idempotent-Replayed: true`; a different body returns 422.",
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : false,
          "schema" : {
            "maxLength" : 255,
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OtpVerify"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TokenResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "`otp.invalid` (wrong or expired code) or `phone.invalid`"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "`account.inactive`: the account is suspended or disabled"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "`idempotency.key-reused`: same Idempotency-Key, different body"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "`otp.attempts-exceeded`: 5 wrong codes; request a new one"
          }
        },
        "summary" : "Verify the code and start a session",
        "tags" : [ "Auth" ]
      }
    },
    "/api/v1/auth/refresh" : {
      "post" : {
        "description" : "Returns new tokens; the presented refresh token stops working. Presenting an already-rotated token revokes the whole session (theft detection), so always send an Idempotency-Key and retry with the same key.",
        "operationId" : "refresh",
        "parameters" : [ {
          "description" : "Language of error messages and SMS: `km` (default) or `en`",
          "in" : "header",
          "name" : "Accept-Language",
          "required" : false,
          "schema" : {
            "enum" : [ "km", "en" ],
            "type" : "string"
          }
        }, {
          "description" : "Optional, 1–255 chars (a UUID is ideal). Retrying with the same key and body replays the first response with `Idempotent-Replayed: true`; a different body returns 422.",
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : false,
          "schema" : {
            "maxLength" : 255,
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TokenResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Invalid request (`code` explains why)"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "`refresh-token.invalid`: unknown, expired, rotated or revoked"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "`idempotency.key-reused`: same Idempotency-Key, different body"
          }
        },
        "summary" : "Rotate the refresh token",
        "tags" : [ "Auth" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "This server",
    "url" : "/"
  } ],
...
}